chat message

+++ user-service/src/main/java/com/socialnetworking/userservice/dto/request/UserEditRequest.java
@@ -1,7 +1,6 @@
 package com.socialnetworking.userservice.dto.request;
 
 import lombok.Data;
-import org.springframework.web.multipart.MultipartFile;
 
 import java.util.Date;
 
@@ -13,5 +12,5 @@ public class UserEditRequest {
     private String username;
     private String bio;
     private Date birthday;
-    private Integer genderId;
+    private Long genderId;
 }
